TallSpinner class separated
